# docs/backend/nivesh-app-openapi.yaml
openapi: 3.1.0
info:
  title: Nivesh App API
  version: 2.0.0
  description: |
    REST API for the Nivesh.ai application — portfolio management, goal planning,
    AI copilot, broker integration, MFD workspace, and admin controls.

    ## Authentication

    All endpoints (except `/api/auth/*`, `/api/public/*`, and `/api/cas-parser-provider/active`)
    require a session cookie set by `POST /api/auth/google`.

    Admin endpoints additionally require the requesting user to have `is_admin = true`.

    ## Base URL

    - **Production**: `https://niveshcopilot.com`
    - **Staging**: `https://staging.niveshcopilot.com`

  contact:
    name: Nivesh
    url: https://nivesh.com
servers:
  - url: https://niveshcopilot.com
    description: Production
  - url: https://staging.niveshcopilot.com
    description: Staging
...
components:
  securitySchemes:
    SessionCookie:
      type: apiKey
      in: cookie
      name: session
      description: Session cookie set by `POST /api/auth/google`.
...
security:
  - SessionCookie: []
paths:
...
  /api/gmail/connect:
    get:
      tags: [gmail]
      operationId: gmail_connect
      summary: Start Gmail OAuth — returns auth_url
      parameters:
        - name: return_to
          in: query
          schema:
            type: string
            nullable: true
      responses:
        "200":
          description: Auth URL
          content:
            application/json:
              schema:
                type: object
                properties:
                  auth_url:
                    type: string
...